Move files and extract crash reporting setup code
