+++ Wire-iOS Share Extension/PreprocessorHelper.h
