+++ Wire-iOS Share Extension/PreprocessorHelper.m
